Fix `ReferenceError: isUndefined is not defined` error
When using the `aws-iot-device-sdk` from within a `create-react-app` (utilizing `babel` and `webpack`), calling `client.device({})` throws the following error:

`ReferenceError: isUndefined is not defined`

This is related to the line at the top of `device/index.js`: `const isUndefined = require('../common/lib/is-undefined')`.

Changing this `const` to `var` solves the issue, but raises the same issue for `mqtt`. Therefore, this commit changes all of the `const` import statements to `var`.