Add some unit test coverage for isUndefined
Albeit trivial, I was working off of an older fork and found
the incorrect `null` check bug in isUndefined. Noticed that
it has since been fixed in master but I already wrote the
unit tests. So we might as well add them. Would have prevented
this kind of bug in the first place.